Trim code duplication & improve project organization
