+++ Cargo.toml
@@ -17,6 +17,7 @@ strip-ansi-escapes = "0.1.0"
 serde = { version = "1.0.*", features = ["derive"] }
 serde_json = "1.0"
 lru-cache = "0.1"
+async-trait = "0.1"
 
 #dbl
 dbl-rs = "0.3"
