+++ src/apis/mod.rs
@@ -1,1 +1,3 @@
 pub mod dbl;
+pub mod wandbox;
+pub mod godbolt;
