+++ src/commands/invite.rs
@@ -4,15 +4,15 @@ use serenity::prelude::*;
 
 use std::env;
 
-use crate::utls::discordhelpers;
+use crate::utls::discordhelpers::embeds;
 
 #[command]
 pub async fn invite(ctx: &Context, msg: &Message, _: Args) -> CommandResult {
     let invite = env::var("INVITE_LINK").expect("Expected invite link envvar");
 
-    let emb = discordhelpers::build_invite_embed(&invite);
+    let emb = embeds::build_invite_embed(&invite);
 
-    let mut emb_msg = discordhelpers::embed_message(emb);
+    let mut emb_msg = embeds::embed_message(emb);
     msg.channel_id.send_message(&ctx.http, |_| &mut emb_msg).await?;
 
     Ok(())
