+++ src/utls/constants.rs
@@ -1,7 +1,7 @@
 //pub static COLOR_OKAY : i32 = 0x046604;
-pub static COLOR_OKAY: i32 = 0x5dbcd2;
+pub static COLOR_OKAY: u32 = 0x5dbcd2;
 //pub static COLOR_FAIL : i32 = 0x660404;
-pub static COLOR_FAIL: i32 = 0xff7761;
+pub static COLOR_FAIL: u32 = 0xff7761;
 pub static ICON_FAIL: &str = "https://i.imgur.com/LxxYrFj.png";
 pub static ICON_VOTE: &str = "https://i.imgur.com/VXbdwSQ.png";
 pub static ICON_HELP: &str = "https://i.imgur.com/TNzxfMB.png";
